[feature/PLAYER-542]: base-components
