+++ js/components/controlBar.js
@@ -14,7 +14,7 @@ var React = require('react'),
   VideoQualityPanel = require('./videoQualityPanel'),
   ClosedCaptionPopover = require('./closed-caption/closedCaptionPopover'),
   MultiAudioPopover = require('./multi-audio/multiAudioPopover'),
-  CloseCaptionMultiAudioMenu = require('./closeCaptionMultiAudioMenu/closeCaptionMultiAudioMenu'),
+  CloseCaptionMultiAudioMenu = require('./close-caption-multi-audio-menu/closeCaptionMultiAudioMenu'),
   Logo = require('./logo'),
   Icon = require('./icon'),
   Tooltip = require('./tooltip');
