+++ js/skin.js
@@ -9,7 +9,7 @@ var React = require('react'),
     ClosedCaptionPanel = require('./components/closed-caption/closedCaptionPanel'),
     DiscoveryPanel = require('./components/discoveryPanel'),
     VideoQualityPanel = require('./components/videoQualityPanel'),
-    CloseCaptionMultiAudioMenu = require('./components/closeCaptionMultiAudioMenu/closeCaptionMultiAudioMenu'),
+    CloseCaptionMultiAudioMenu = require('./components/close-caption-multi-audio-menu/closeCaptionMultiAudioMenu'),
     SharePanel = require('./components/sharePanel'),
     MoreOptionsPanel = require('./components/moreOptionsPanel'),
     AdScreen = require('./views/adScreen'),
